translate-web-app/backend/requirements.txt
# Core Google Generative AI (Gemini API)
google-genai>=0.2.0
pydantic>=2.0.0

# Document processing libraries
python-pptx>=0.6.21
python-docx>=1.0.0
pymupdf>=1.23.0
pdfplumber>=0.10.0
reportlab>=4.0.0

# PDF Layout Detection (for advanced PDF translation)
opencv-python-headless>=4.8.0
numpy>=1.24.0
onnx>=1.15.0
onnxruntime>=1.16.0
huggingface_hub>=0.19.0
tqdm>=4.66.0

# Optional but recommended dependencies
chardet>=5.0.0
Pillow>=10.0.0

# Environment and configuration
python-dotenv>=1.0.0
